\name{pvxmlrpclibs-package}
\alias{pvxmlrpclibs-package}
\alias{pvxmlrpclibs}
\docType{package}
\title{
BridgeDB libraries.
}
\description{
Package with the BridgeDB libraries, mostly aimed for use
in the bridgedbr package. The BridgeDB librarys packaged in this
package are de written by: Martijn van Iersel, Isaac Ho, Jianjiong Gao and Alex Pico, and are licensed XXXXXX.
}
\author{
Christ Leemans

Maintainer: Christ Leemans <christleemans@gmail.com>
}
\keyword{ package }
\keyword{ BridgeDB }
\keyword{ identifier mapping }
\keyword{ WikiPathways }
